Implement robust GND10 reader.
This implements a class usable from elsewhere for reading GND10
messages.

The new class will try to reset and retry in case of a USB disconnect or serial
port read timeout.

I've seen cases before where the GND10 needed a hard manual USB disconnect to
get going again, but I haven't been able to reproduce it lately. The corrective
action in case the reader does not know of the failure mode, is "reset the
power" or "yank out and reinsert the usb cable". The goal is that the software
should self heal after that.