Flaky input should not be logged as warnings.
The input from GND10 often contains garbled data, which we try to detect
as new and interesting frame types. This leads to a very noisy log,
where important messages are lost in all the other output.

Since the decoder has a workable understanding of FDX now,
stop treating these transfer errors as errors to be logged.