Add two long conf_ messages.
Based on the input from Bob Baker, I believe this has something
to do with Angle corrections and Speed corrections to the NX2
server.

Without further data and dumps it is hard to say, and I don't have a windows
computer/vm at hand to do this testing myself right now.