Add a tokenizer that uses frame length.
Today I realized that the second octet in the 3 octet header
is in fact the message length.

This has been bugging me forever, since it made no sense to have
variable message lengths without a frame length field.

Since it is off by one octet, I suspect that either the first or last PDU octet
is special. Need to look at the data for that, but at least this is progress.